Made some layout changes to the directory structure.
The executables are now in the base64 directory.

There is a new base64 program, which can encode and decode.
This deprecates the older b64enc, b64dec, encoder and decoder programs.

The base64 program reads and writes to files that it opens (instead of STDIN and STDOUT),
 which also makes it easier for the coder (ie. me) to ensure that the files are opened
 in binary mode - this is necessary because some irritating OS'es out there require
 this mode to deal with non-text characters in any file stream. Sigh...

Updated the docs and added a BENCHMARKS file, with some interesting statistics.